Refactor: 통합 타입 시스템 적용 및 코드 정리 - participate 컴포넌트 타입 정의를 src/types/components/participate.ts로 통합 - 불필요한 더미 데이터 파일(meetingMaterialsData.ts) 제거 - 네비게이션 함수 인라인화로 코드 간소화 - 사용하지 않는 import 및 타입 정의 정리 - 빌드 에러 해결 및 코드 일관성 개선
