--- src/components/participate/data/meetingMaterialsData.ts
@@ -1,14 +1,0 @@
-import { type BoardItem } from '@/types/components';
-
-export const meetingMaterialsData: BoardItem[] = [
-  { id: 10, title: '[정기회의] 제10회 정기회의 자료입니다.', date: '2025.07.07' },
-  { id: 9, title: '[정기회의] 제9회 정기회의 자료입니다.', date: '2025.07.07' },
-  { id: 8, title: '[정기회의] 제8회 정기회의 자료입니다.', date: '2025.07.07' },
-  { id: 7, title: '[정기회의] 제7회 정기회의 자료입니다.', date: '2025.07.07' },
-  { id: 6, title: '[정기회의] 제6회 정기회의 자료입니다.', date: '2025.07.07' },
-  { id: 5, title: '[정기회의] 제5회 정기회의 자료입니다.', date: '2025.07.07' },
-  { id: 4, title: '[정기회의] 제4회 정기회의 자료입니다.', date: '2025.07.07' },
-  { id: 3, title: '[정기회의] 제3회 정기회의 자료입니다.', date: '2025.07.07' },
-  { id: 2, title: '[정기회의] 제2회 정기회의 자료입니다.', date: '2025.07.07' },
-  { id: 1, title: '[정기회의] 제1회 정기회의 자료입니다.', date: '2025.07.07' },
-];
