--- src/components/participate/data/types.ts
@@ -1,5 +1,0 @@
-export interface FAQItem {
-  id: number;
-  question: string;
-  answer: string;
-}
