+++ src/types/components/participate.ts
@@ -1,13 +1,8 @@
-// 참여하기 관련 컴포넌트 Props 타입
-
 /**
  * FAQ 섹션 컴포넌트 Props
  */
 export interface FAQSectionProps {
-  faqs: Array<{
-    question: string;
-    answer: string;
-  }>;
+  faqData: FAQItem[];
   className?: string;
 }
 
@@ -19,4 +14,50 @@ export interface VolunteerApplicationModalProps {
   onClose: () => void;
   onSubmit: (data: any) => void;
   loading?: boolean;
+}
+
+/**
+ * FAQ 아이템 타입
+ */
+export interface FAQItem {
+  id: number;
+  question: string;
+  answer: string;
+}
+
+/**
+ * 후원금 사용처 아이템 타입
+ */
+export interface SponsorshipUsageItem {
+  icon: any;
+  title: string;
+  description: string;
+}
+
+/**
+ * 후원 방법 아이템 타입
+ */
+export interface SponsorshipMethodItem {
+  icon: any;
+  title: string;
+  description: string;
+}
+
+/**
+ * 자원봉사 신청 폼 데이터 타입
+ */
+export interface VolunteerFormData {
+  name: string;
+  birthDate: string;
+  gender: string;
+  phone: string;
+  email: string;
+  emailDomain: string;
+  address: string;
+  activityFields: string[];
+  availableTimes: string[];
+  participationPeriod: string;
+  hasExperience: string;
+  motivation: string;
+  privacyConsent: boolean;
 }
